pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["tyreln <tyrelnarciso@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.8.0"
djoser = "^2.1.0"
Pillow = "^8.3.2"
django-environ = "^0.8.0"
django-cloudinary-storage = "^0.3.0"
django-cleanup = "^5.2.0"
psycopg2 = "^2.9.2"
django-filter = "^21.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
psycopg2-binary = "^2.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
